requirements.txt
scrapy
scrapy-splash
pillow
boto
shub
pip install scrapy-crawlera
numpy
selenium
json
pickle
pandas